fix compile issues

+++ include/ccapi_cpp/service/ccapi_execution_management_service_cryptocom.h
@@ -204,7 +204,7 @@ class ExecutionManagementServiceCryptocom : public ExecutionManagementService {
   }
 
   void convertRequestForWebsocket(rj::Document& document, rj::Document::AllocatorType& allocator, const WsConnection& wsConnection, const Request& request,
-                                  int wsRequestId, const TimePoint& now, const std::string& symbolId,
+                                  unsigned long wsRequestId, const TimePoint& now, const std::string& symbolId,
                                   const std::map<std::string, std::string>& credential) override {
     switch (request.getOperation()) {
       case Request::Operation::CREATE_ORDER: {
